Accept rush projects instead of rush json as action input, refactor structure
