--- __mocks__/rush.json
@@ -1,21 +1,0 @@
-// Comments should be ignored
-{
-  "projects": [
-    {
-      "packageName": "foo",
-      "projectFolder": "__mocks__/packages/foo"
-    },
-    {
-      "packageName": "goo",
-      "projectFolder": "__mocks__/packages/goo"
-    },
-    {
-      "packageName": "bar",
-      "projectFolder": "__mocks__/packages/bar"
-    },
-    {
-      "packageName": "baz",
-      "projectFolder": "__mocks__/packages/baz"
-    }
-  ]
-}
