--- __mocks__/rush.ts
@@ -1,20 +1,0 @@
-export default {
-  projects: [
-    {
-      packageName: 'foo',
-      projectFolder: '__mocks__/packages/foo',
-    },
-    {
-      packageName: 'goo',
-      projectFolder: '__mocks__/packages/goo',
-    },
-    {
-      packageName: 'bar',
-      projectFolder: '__mocks__/packages/bar',
-    },
-    {
-      packageName: 'baz',
-      projectFolder: '__mocks__/packages/baz',
-    },
-  ],
-};
