+++ __mocks__/rushProjects.json
@@ -1,0 +1,18 @@
+[
+  {
+    "packageName": "foo",
+    "projectFolder": "__mocks__/packages/foo"
+  },
+  {
+    "packageName": "goo",
+    "projectFolder": "__mocks__/packages/goo"
+  },
+  {
+    "packageName": "bar",
+    "projectFolder": "__mocks__/packages/bar"
+  },
+  {
+    "packageName": "baz",
+    "projectFolder": "__mocks__/packages/baz"
+  }
+]
