+++ global.d.ts
@@ -1,4 +1,4 @@
-interface RushPackage {
+interface RushProject {
   packageName: string;
   projectFolder: string;
 }
