--- jest.config.js
@@ -1,8 +1,0 @@
-module.exports = {
-  clearMocks: true,
-  restoreMocks: true,
-  preset: 'jest-preset-typescript',
-  testEnvironment: 'node',
-  collectCoverage: true,
-  coverageReporters: ['text-summary', 'text'],
-};
