+++ src/github-action.ts
@@ -1,45 +1,35 @@
+import { debug, getInput, setFailed, setOutput } from '@actions/core';
 import {
-  getAllPackages,
-  getChangedPackages,
+  getPullRequestNumber,
+  isMainBranch,
   getDiffTargetMain,
   getDiffTargetPullRequest,
   getTagForDeployment,
-  readJson,
-} from './utils';
-import { debug, getInput, info, setFailed, setOutput } from '@actions/core';
-import { getPullRequestNumber, isMainBranch } from './github';
+} from '@src/util/github-context';
+import { findChangedProjects } from '@src/lib/find-changed-projects';
 
-const run = (): void => {
+export const run = (): void => {
   try {
-    const environment = getInput('environment');
-    const rushJsonPath = getInput('rushJsonPath');
-
     const pullRequestNumber = getPullRequestNumber();
     const isMain = isMainBranch();
-
     if (!pullRequestNumber && !isMain) {
       throw new Error('This action only supports push event on main branch or pull request events');
     }
-
+    const environment = getInput('environment');
+    const rushProjectsInput = getInput('rushProjects');
     const tagForDeployment = getTagForDeployment(pullRequestNumber, environment);
     const diffTarget = isMain
       ? getDiffTargetMain(tagForDeployment)
       : getDiffTargetPullRequest(tagForDeployment);
-    const rushPackages: RushPackage[] = readJson(rushJsonPath).projects;
-
-    debug(JSON.stringify(rushPackages, null, 2));
-
+    const rushProjects: RushProject[] = JSON.parse(rushProjectsInput);
+    debug(JSON.stringify(rushProjects, null, 2));
     const changedProjects = diffTarget
-      ? getChangedPackages(diffTarget, rushPackages)
-      : getAllPackages(rushPackages);
-
-    info(JSON.stringify(changedProjects, null, 2));
-
+      ? findChangedProjects(diffTarget, rushProjects)
+      : rushProjects;
+    debug(JSON.stringify(changedProjects, null, 2));
     setOutput('changedProjects', changedProjects);
     setOutput('tag', tagForDeployment);
-  } catch (e) {
-    setFailed(e.message);
+  } catch (error: any) {
+    setFailed(error.message);
   }
 };
-
-run();
