+++ src/lib/find-changed-projects/__mocks__/rushProjects.ts
@@ -1,0 +1,18 @@
+export const rushProjects = [
+  {
+    packageName: 'foo',
+    projectFolder: '__mocks__/packages/foo',
+  },
+  {
+    packageName: 'goo',
+    projectFolder: '__mocks__/packages/goo',
+  },
+  {
+    packageName: 'bar',
+    projectFolder: '__mocks__/packages/bar',
+  },
+  {
+    packageName: 'baz',
+    projectFolder: '__mocks__/packages/baz',
+  },
+]
