+++ src/lib/find-changed-projects/find-changed-projects.ts
@@ -1,0 +1,13 @@
+import { isChangeInPath } from "@src/util/git-client";
+
+export const findChangedProjects = (
+  diffTarget: string,
+  rushProjects: RushProject[]
+) => {
+  return rushProjects.reduce<RushProject[]>((changes, project) => {
+    if (isChangeInPath(diffTarget, project.projectFolder)) {
+      changes.push(project)
+    }
+    return changes;
+  }, []);
+};
