+++ src/lib/find-changed-projects/index.ts
@@ -1,0 +1,1 @@
+export * from './find-changed-projects';
